speedup and fix audio/subtitles switch
When switching audio/subtitles then there is a need to flush buffers. This was done by seeking 1s back, which was not optimal. It caused total hang when start position was wrong (simulated on some live dash streams with fmp4) and also seeking backward is problematic. This small change will set seek to 0s, so no seeking at all, but buffers will be still flushed.